+++ src/raindrop_todoist_syncer/main.py
@@ -9,7 +9,7 @@ from raindrop_todoist_syncer.rd_process import RaindropsProcessor
 from raindrop_todoist_syncer.rd_token import RaindropAccessTokenRefresher
 from raindrop_todoist_syncer.rd_client import RaindropClient
 
-from raindrop_todoist_syncer.td_task_create import TodoistTaskCreator
+from raindrop_todoist_syncer.td_task import TodoistTaskCreator
 
 from raindrop_todoist_syncer.logging_config import configure_logging
